backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
PyPDF2
openai
requests
sentence-transformers
psycopg2-binary
pgvector